requirements.txt
farm-haystack[inference]
farm-haystack[sql]
farm-haystack[elastic]
